+++ src/actions/hashtags.ts
@@ -1,6 +1,17 @@
 import { ActionError, defineAction } from "astro:actions";
 import { isAuthorized } from "./_helpers";
-import { db, SharedTag, eq, and, User, or } from "astro:db";
+import {
+  db,
+  SharedTag,
+  eq,
+  and,
+  User,
+  or,
+  Todo,
+  count,
+  not,
+  like,
+} from "astro:db";
 
 import { v4 as uuid } from "uuid";
 import { z } from "zod";
@@ -28,6 +39,54 @@ export const getSharedTags = defineAction({
       sharedToUser,
       sharedByUser,
     };
+  },
+});
+
+export const getHashtags = defineAction({
+  handler: async (_, c) => {
+    const userId = isAuthorized(c).id;
+    const hashtags = await db
+      .select({ text: Todo.text })
+      .from(Todo)
+      .where(
+        and(
+          eq(Todo.isDeleted, false),
+          eq(Todo.userId, userId),
+          like(Todo.text, "%#%"),
+        ),
+      )
+      .then((rows) => rows.map((row) => row.text))
+      .then((texts) =>
+        texts.reduce((acc, val) => {
+          const matches = val.match(/#[a-zA-Z0-9]+/g);
+          if (matches) {
+            matches.forEach((match) => {
+              acc.add(match);
+            });
+          }
+          return acc;
+        }, new Set<string>()),
+      )
+      .then((set) => Array.from(set))
+      .then((arr) => arr.map((text) => text.replace("#", "")));
+
+    const untaggedTodos = await db
+      .select({ count: count(Todo.id) })
+      .from(Todo)
+      .where(
+        and(
+          eq(Todo.isDeleted, false),
+          eq(Todo.userId, userId),
+          not(like(Todo.text, "%#%")),
+        ),
+      )
+      .then((rows) => rows[0].count);
+
+    if (untaggedTodos > 0) {
+      hashtags.unshift("~");
+    }
+
+    return hashtags;
   },
 });
